socket: Handle FIN by throwing an exception after processing the payload.
Okay, I had it wrong last time.  We need to be able to signal the client
about receiving a FIN even if there is no payload to process and we will
add such a notification in the next commit.  By processing the FIN
transition first, we would end up signaling the client that we've
received FIN (and hence to expect no more data)... and then we'd deliver
more data if there was a payload.  Not really acceptable.  So, we have
to do the FIN processing after delivering the last payload to the
client.  The easiest way to do this is just throw an exception at the
end of processing the FIN packet.