socket: Add tcp::socket_observer::socket_fin_recvd().
This notifies the client that we've received the FIN sequence number for
this socket - the remote guy is done transmitting data to us.  We should
process the remainder of the receive queue as necessary and if we're in
a passive close state (i.e. we haven't issued close_send() yet) then the
client eventually needs to close our end so we send the FIN and clean
up the connection.