socket: Use different bounds for valid ACK checking in the SYN_SENT states.
According to the spec, it is illegal for the remote guy to send us an
ACK with number ISS (this would effectively ACK an empty window before
the initial SYN octet - I personally think that's fine and it would have
not state change, but the TCP spec says to drop it or possibly send a
RST).  In the initial SYN_SENT states, before we've received the first
ACK we have SND_UNA=ISS so the bound check was not quite strict enough.
Probably harmless, but let's adhere to the spec.

This also fixes the one failing blackbox test.