socket: Fix out-of-order acceptable packet handling.
If a packet lies (fully or partially) inside the receive window, it's
classified as an acceptable packet.  However, that doesn't mean it is
the NEXT expected packet - it could be out of order.  We were
erroneously just appending every acceptable packet to the receive queue
when in fact we should probably be appending these to a secondary
out-of-order queue for further processing.

For now, we just drop them and rely on the remote end to retransmit.