tcp_blackbox_test: Add test demonstrating that we transmit an extra ACK if we consume the rx queue in the rx_append() callback context.
We do the rx_append() callback first, in which the client can do a
read().  When the client does a read() this expands the available
receive window so we post an ack to notify the remote guy.  Then, we
return back to process_payload_synchronized() and send another ACK which
ends up being a duplicate.  Duplicate ACKs are bad since they will
confuse the congestion control protocol on the other end.