socket: Remove post_ack() parameters.
We always post the latest information.  In the future we might even want
to throttle the information we post (i.e. defer window size updates so
that we don't give a bunch of tiny updates to the remote end and cause
him to send a bunch of tiny packets instead of full-MSS ones) and it
will be much easier to do that all in one place.