socket: Differentiate between types of ACK and don't transmit duplicate update ACKs.
There are two types of ACK: ACKs that are sent in response to a local
state change updating the remote guy that either our receive window
size changed or we received new data from him, and ACKs that are sent in
response to unexpected packets we've received from the network.

The unexpected packets will be packets that are out-of-range for some
reason, probably because of dropped packets on the network.  If these
packets trigger an ACK we must always transmit it because the remote guy
may not have seen the original one.

However, for update packets we definitely don't want to send duplicate
ACKs.  One example is where we receive a packet, get the client to
process it in-line, the client reads() all of the data and we post an
update ACK for that, and then finally we post an update ACK because we
received data from the remote guy.  These two ACKs end up being
duplicates and we want to merge them into one.  It's important that we
do this because duplicate ACKs are used by some congestion control
algorithms and we don't want to confuse them.