kern: Move remaining top-level lokios.1 sources into new kern/ module.
This cleans things up significantly, but required a lot of #include and
module.mk adjustments.  We also had to change the lokios.1 module.mk to
know that libsupc++ is a third-party library that provides weak symbols
that we want to override; we can't just include it in the \( ... \)
libaries passed to the linker because then it's weak symbols can end up
overriding our symbols (especially for 'operator new') and everythign
blows up.  So, make sure to define our libraries first and libraries
that provide weak symbols last.