libc: Provide working implementations of malloc() and free().
Sigh.  The C++ exception-handling stuff needs to have working versions
of malloc() and free().  The first time you throw an exception, the
libraries go and parse a bunch of exception-handling tables into sorted
data structures that are fast to traverse.  This involves calling
malloc().  The size of the allocated buffer is larger than 4K, so a
simple page_alloc doesn't work.

We *can* return NULL, but then the exception-handling library will fall
back to a slow linear search when throwing an exception - probably
unacceptable for us.

Finally, when throwing an exception, the memory for the exception is
allocated on the heap.  Sad but true.  We were handling this before via
the page_alloc case and continue to do so now with the buddy allocator.

It's unclear who else might be using malloc (__cxa_demangle is surely a
a candidate), so... I give up and just implement it.