page: Handle discontiguous, unsorted e820 regions below 32M and eliminate the sbrk region.
Now that the sbrk() API is gone, we can relax some of our low-mem area
requirements.  Thie buddy allocator bitmap now starts immediately after
the end of the kernel image and memory in the page immediately after the
end of the bitmap is the first page we will add to the buddy allocator.
This frees up most of the memory that was reserved for sbrk for anybody
to use.  It also means that we can iterate over all the low-mem E820
regions and free them on a page-by-page basis even if they aren't all
contiguous.

All this was made possible by implementing malloc/free in terms of the
buddy allocator.  malloc/free are there just for the external
exception-handling and name-demangling stuff.  So much junk just to
support that... at least it is a lot cleaner now.