page: Use a local_vector<> instead of a regular one when populating the high-mem buddy pages.
We used a local_vector<> for this exact same data structure when doing
the low-mem stuff, so we may as well continue to do it in the high-mem
code as well.  This has the added effect of not gobbling up extra buddy
pages while we are populating the buddy allocator - although there's
technically nothing wrong with doing that and the mmap stuff is going to
do it if it needs more page-table pages anyways.