socket: Use the ack queue to ensure we complete all send ops in order.
Under some weird conditions (notably if the remote guys shrinks his
receive window and we are doing retransmits), it would be possible for a
high-numbered ACK packet to trigger client send completions near the end
of the list of unacked ops if the send ops near the front of the list
still have posted tx_ops to the chip.  The scenario would be something
like this:

1. We post ops 1 thru 10 and these all complete.
2. The remote guy shrinks his receive window.
3. Our transmit timer fires.
4. We post as many ops as we can in the new window, say ops 1 thru 5,
   but the posted ops don't complete just yet - queued in the NIC.
5. We immediately receive an ACK for ops 1 thru 10.

At this point, ops 1-5 would have had positive refcounts so we wouldn't
complete them.  However ops 6-10 would have zero refcounts and the code
would happily go and complete them.  But from the client's perspective
he'd see out-of-order completions.  Yuck!

So, we actually have to delay the send completions for ops 6-10 until we
have tx_op completions for the (now-redundant) stuff still on the chip
queue.  We can't just complete ops 1-5 and ignore the tx_op chip
completions because those tx_ops are referring to dma_alps in the
send_ops and if we complete the send_ops then the client could reuse
those buffers for other data.  The DMA addresses would remain valid but
we'd be transmitting completely different data on the wire - potentially
private data.

Ugly.