socket: Transition through TCP_DRAIN_NIC_OPS on the reset path.
When we get a RST=1 packet, we may have a bunch of stuff queued up.
Some in the chip as tx_ops and some in socket queues as send_ops.
Before we allow the socket to be deleted, we need to wait for all the
chip tx_ops to complete and then we also need to fail any remaining
send_ops back up to the client once there are no longer any tx_ops
referencing them.

The current TCP_DRAIN_NIC_OPS code is for a graceful shutdown where
we've received an ACK for our FIN, so the send_ops queues should be
empty at this point.  But on the reset path that isn't necessarily the
case, so we exand socket_closed() handling to fail any send_ops that may
still be queued up on the socket when we get into TCP_CLOSED.

This also means that on the reset path we will get the following
notifications:

    observer->socket_reset
        - the socket has been reset, please don't post any more send_ops
          but also please don't delete us yet
    observer->socket_closed
        - the socket has entered the TCP_CLOSED state, all posted
          send_ops have been failed and we surely aren't going to
          receive any more packets - delete us at your leisure