socket: Add TCP_DRAIN_NIC_OPS state.
This state is on the normal socket closure path.  We've received an ACK
for our FIN and completed the FIN send_op.  This means we've also
completed all prior send_ops.  However, there may still be posted tx_ops
on the chip queue for non-send_op related transmissions such as regular
ACKs.  We need to wait for chip completions for that stuff before we can
transition all the way to TCP_CLOSED and deleted.