socket: On the reset path, wait for all tx_ops to complete and then fail all send_ops.
When we get a RST=1 packet, we may have a bunch of stuff queued up.
Some in the chip as tx_ops and some in socket queues as send_ops.
Before we allow the socket to be deleted, we need to wait for all the
chip tx_ops to complete and then we also need to fail any remaining
send_ops back up to the client once there are no longer any tx_ops
referencing them.