socket: Provide a destructor that will clean up any pages left on the receive queue.
If the socket closes in error, say due to a reset, it's perfectly
acceptable for the client to not consume the entire receive queue.  In
fact, it's perfectly acceptable even if the socket closes gracefully.

So, we add a socket destructor and clean up any leftover pages.  This
has a side-effect: the hash table class also cleans up after itself if
there are nodes leftover in the hash table and it will call their
destructors.  The DHCP test doesn't create any sockets, but it does have
a net::interface and that does have a socket hash table and the socket
hash table destructor does get called.  The compiler has no way of
knowing that the hash table will always be empty, so it has to emit the
hash table code that will call the socket destructors for whatever
happens to be queued in the hash table - even though it's always going
to be empty.

Short version: this means we need a mock socket destructor method which
we can get by linking against msocket.