socket: Defer observer->socket_closed callback to a cpu work entry.
When we do the socket_closed callback, the client is perfect within
their rights to call tcp_delete() and nuke our socket into oblivion.
But our socket_closed could be in response to an ACK in the LAST_ACK
state, which means we're deep in the middle of processing the FIN at the
end of the ACK queue.  In particular, the send_op for the FIN won't have
been freed up yet (since we're in the midst of invoking its callback) so
there will be at least one send_ops_slab element in use and a tcp_delete
would die.  Not to mention the fact that we will want to continue
iterating over the (now-empty) send ops queue before returning.

Transition immediately, but defer the notification to a safer time and
place.