schedule: Assert that a timer is not armed when it is being destroyed.
We were relying on the kdlink destructor to detect if we were destroying
an armed timer, however for long-duration timers they'll be populated in
the timer min-heap instead of in the timer wheel so their kdlink won't
be used.  So, now we also assert that we aren't in the min-heap either.

This causes many unittests to fail.