pci: Provide an unclaimed_driver pci driver to claim devices no other driver wants.
The way the probing stuff works is that a pci::dev only exists if some
driver somewhere claims it.  In the future we want to support wapi
queries of the PCI hierarchy, even for devices that aren't claimed by
any driver.  So, we create a dummy driver that will match against
anything with the lowest possible non-zero score (1).  It will
instantiate pci::devs via a slab so they shouldn't occupy a ton of
space.