elf_image: Allow ELFOSABI_GNU images.
The GNU ELF Extensions seem to provide for a few things beyond the base
ELF spec.  One of them is relocations that are computed at runtime
(blech - we don't support this) and another is tagging a symbol so that
the dynamic linker ensures that it's globally unique within a process.
Template static members are an example of this and we'll start using
those shortly.

We don't actually have to do anything since we're the only binary in the
kernel process image, so any symbols are already guaranteed to be
unique.  However, the ELF binary will now be tagged differently, so we
just need to recognize that fact in the bootloader.