http: Add http::request class that parses an inbound HTTP request.
This extracts header fields into a hash table object, finds the message
body and reads it into a kernel::string object.  It also decodes the
method.

This parser only works for request bodies that contain a
"Content-Length" header.  It does not support the chunked encoding or
any other type of transfer encoding such as gzipped data.