delegate: Make function delegates slightly more efficient.
One of the more recent C++ changes was to allow static functions to be
used as template arguments - external linkage is no longer required to
use a function as a template argument.

This change updates the fbounce() method to become a template method
that takes a function template argument and invokes it.  Since the
function is now a template argument instead of a passed-in pointer, we
can branch directly to the target function instead of doing an indirect
jump via the function pointer.  Unfortunately we still need to strip the
(now-junk) pointer from the argument list since we need to maintain the
same signature that we use for method delegates where the first argument
is an object pointer.  Still, this is an improvement.