pci: Don't emit the domain as part of a pci device string.
The domain is already encoded in the wapi path, so don't re-encode it in
the name.  This will make the client's life easier when we start
attaching pci::dev objects below the domain in the wapi hierarchy - the
client won't have to strip the domain out of the slot name.