platform: Rename reboot_guest -> reboot.
exit_guest() exists to shut down either real hardware or a VM and in the
case of a VM to shutdown with a particular exit code.  This is the only
interesting operation that is specific to a VM.

reboot_guest() doesn't make sense as a VM operation since it isn't doing
anything that is specific only to a VM.  We're just rebooting the kernel
and all APs here; that's a regular hardware operation not a VM
operation.

Rename this to avoid any other confusion about its intent.