http: Add method to check if a request came with a JSON body.
This also assumes that if we're asking the question "is there a JSON
body" that you care about being able to actually identify the body type.
So, if there is a body but no Content-Type field to tell us what is,
rather than returning false (hey, it *could* be a JSON body and somebody
just forgot to put in the Content-Type header, so returning false
wouldn't always be the right choice) throw an exception.