acpi: Print json oem table ids as a string instead of each char.
This is probably still bad, but it works here.  Apple has a '\0' at the
end of one of their table ids and we were returning that in the json
output, causing python requests to throw an exception.

We should probably just give up and go to hex, or maybe do something a
bit smarter with an isascii() test and checks for trailing NULs, but...
for now just do a %.6s printf field.