wapi: Provide wapi::global_node which registers a wapi node for deferred linking into the wapi tree.
Because of the initialization order fiasco, we can't have global
object constructors manipulating fields of other globals since those
other globals may not have been initialized yet.  So we can't just build
the wapi tree at global ctor time.  However, if objects record their
parent nodes when they are constructed and then queue themselves up on a
(compile-time-inited) klist<> then we can walk that list later on and
get each node to register itself with it's
now-guaranteed-to-be-constructed parent node.