buddy_allocator: Change page-population API so we can count the total number of pages given to the allocator.
The buddy allocator has a start and an end address but there may be
holes in the memory contained in that range that are not suitable for
allocation purposes.  For this reason, we can't get a good fix on the
number of pages that are actually available for use.  When we are
populating the allocator, we are simply calling the buddy_free() method
which doesn't really help us count the number of available pages.  By
adding a buddy_populate() method, we can do the initial page population
with an API that will count how many pages it has been given.

This, together with the count of free pages that we are already
maintaining on each call to buddy_alloc() or buddy_free(), gives us the
number of allocated pages as well without any extra runtime overhead.